+++ __tests__/pages/ContributorsPage.test.jsx
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "vitest"
 import { screen, render } from "@testing-library/react"
-import ContributorsPage from "./ContributorsPage"
+import ContributorsPage from "../../src/pages/ContributorsPage"
 
 describe("Contributors Page test", () => {
   test("Should render Contributors Page", () => {
