+++ __tests__/pages/HomePage.test.jsx
@@ -1,7 +1,7 @@
 import { describe, expect, test } from "vitest"
 import { screen, render } from "@testing-library/react"
 import { BrowserRouter as Router } from "react-router-dom"
-import HomePage from "./HomePage"
+import HomePage from "../../src/pages/HomePage"
 
 describe("HomePage test", () => {
   test("Should render Home Page", () => {
