+++ contributing.md
@@ -58,13 +58,19 @@ pnpm dev
 git checkout -b [name_of_your_new_branch]
 ```
 
-9. Add your changes.
+9. Make sure all the test cases pass
+
+```
+pnpm test
+```
+
+10. Add your changes.
 
 ```diff
 git add .
 ```
 
-10. Add your commits
+11. Add your commits
 
 ```diff
 git commit -m "Your Changes"
@@ -82,18 +88,18 @@ git commit -m "Your Changes"
 > pnpm format
 > ```
 
-11. Set upstream command
+12. Set upstream command
 
 ```diff
 git remote add upstream https://github.com/priyankarpal/ProjectsHut.git
 ```
 
-12. Push your commits
+13. Push your commits
 
 ```diff
 git push -u origin [Your-branch-name]
 ```
 
-13. Create a PR
+14. Create a PR
 
-14. Wait for merge
+15. Wait for merge
